+++ src/app/advanced-examples/minesweeper/components/controls.tsx
@@ -1,37 +1,36 @@
 import RxFM from "rxfm";
-import { Observable, combineLatest, of, timer } from "rxjs";
+import { Observable, of, timer } from "rxjs";
 import { switchMap, filter, map, scan, startWith, distinctUntilChanged } from "rxjs/operators";
-import { GameStage, CellAction } from "../types";
+import { MinesweeperGameState } from "../game-logic/minesweeper-game";
+import { CellAction } from "../types";
 
 interface ControlsProps {
-  startTime: Observable<number | undefined>;
-  endTime: Observable<number | undefined>;
-  gameStage: Observable<GameStage>;
+  gameState: Observable<MinesweeperGameState>;
   dispatch: (action: CellAction) => void;
 }
 
-export const Controls = ({ startTime, endTime, gameStage, dispatch }: ControlsProps) => {
-  const gameTime = combineLatest([startTime, endTime]).pipe(
-    switchMap(([start, end]) => {
-      if (!start) return of(0);
-      else if (!end) return timer(0, 1000);
-      return of(Math.round((end - start) / 1000));
+export const Controls = ({ gameState, dispatch }: ControlsProps) => {
+  const gameTime = gameState.pipe(
+    switchMap(({ startTime, endTime }) => {
+      if (!startTime) return of(0);
+      else if (!endTime) return timer(0, 1000);
+      return of(Math.round((endTime - startTime) / 1000));
     }),
   );
 
-  const HighScore = combineLatest([startTime, endTime]).pipe(
-    filter(([start, end]) => Boolean(start) && Boolean(end)),
-    map(([start, end]) => Math.round((end! - start!) / 1000)),
+  const HighScore = gameState.pipe(
+    filter(({ startTime, endTime }) => Boolean(startTime) && Boolean(endTime)),
+    map(({ startTime, endTime }) => Math.round((endTime! - startTime!) / 1000)),
     scan((lowestTime, time) => Math.min(lowestTime, time), Infinity),
     startWith(undefined),
     distinctUntilChanged(),
     switchMap(highScore => highScore ? <div>Highscore: {highScore}s</div> : of(null)),
   );
 
-  const WinLoseMessage = gameStage.pipe(
-    switchMap(stage => {
-      if (stage === 'win') return <div>You Win!</div>;
-      if (stage === 'gameOver') return <div>Game Over!</div>;
+  const WinLoseMessage = gameState.pipe(
+    switchMap(({ gameStage }) => {
+      if (gameStage === 'win') return <div>You Win!</div>;
+      if (gameStage === 'gameOver') return <div>Game Over!</div>;
       return of(null);
     }),
   );