+++ src/app/advanced-examples/minesweeper/components/game-cell.tsx
@@ -1,6 +1,6 @@
-import RxFM, { destructure, conditional, FC, and } from "rxfm";
+import RxFM, { FC } from "rxfm";
 import { Observable } from "rxjs";
-import { map } from "rxjs/operators";
+import { map, pluck } from "rxjs/operators";
 import { indexToVector, NEIGHBORS_COLOR_MAP } from "../constants";
 import { MinesweeperCell } from "../game-logic/minesweeper-cell";
 import { CellAction, CellActionType } from "../types";
@@ -12,29 +12,30 @@ interface GameCellProps {
 }
 
 export const GameCell: FC<GameCellProps> = ({ cell, index, dispatch }) => {
-  const { neighbors, hasNeighbors, symbol, color, isCleared, isUndiscovered } = destructure(cell);
 
   const handleCellAction = (type: CellActionType) => index.pipe(
     map(index => () => dispatch({ type, cell: indexToVector(index) }))
   );
 
-  const cellText = conditional({
-    if: and(isCleared, hasNeighbors),
-    then: neighbors,
-    else: symbol,
-  });
+  const cellText = cell.pipe(
+    map(({ isCleared, hasNeighbors, neighbors, symbol }) => isCleared && hasNeighbors ? neighbors : symbol),
+  );
 
   const styles = {
-    backgroundColor: color,
-    color: neighbors.pipe(
-      map(neighbors => NEIGHBORS_COLOR_MAP[neighbors]),
+    backgroundColor: cell.pipe(pluck('color')),
+    color: cell.pipe(
+      map(({ neighbors }) => NEIGHBORS_COLOR_MAP[neighbors]),
     ),
-    fontSize: conditional(symbol, '12px', '14px'),
+    fontSize: cell.pipe(
+      map(({ symbol }) => symbol ? '12px' : '14px'),
+    ),
   };
 
   return <div
     style={styles}
-    class={['minesweeper-cell', conditional(isUndiscovered, 'raised')]}
+    class={['minesweeper-cell', cell.pipe(
+      map(({ isUndiscovered }) => isUndiscovered && 'raised'),
+    )]}
     onClick={handleCellAction('discover')}
     onContextMenu={handleCellAction('flag')}
   >