+++ src/app/advanced-examples/minesweeper/minesweeper.tsx
@@ -1,9 +1,10 @@
-import RxFM, { destructure, FC } from 'rxfm';
+import RxFM, { FC } from 'rxfm';
 import { Subject } from 'rxjs';
 import { minesweeperGameLoop } from './game-logic/minesweeper-game';
 import { CellAction } from './types';
 import { GameBoard } from './components/game-board';
 import { Controls } from './components/controls';
+import { pluck } from 'rxjs/operators';
 
 import './minesweeper-styles.css';
 
@@ -11,10 +12,10 @@ export const Minesweeper: FC = () => {
   const cellAction = new Subject<CellAction>();
   const dispatch = (action: CellAction) => cellAction.next(action);
 
-  const { board, startTime, endTime, gameStage } = destructure(minesweeperGameLoop(cellAction));
+  const gameState = minesweeperGameLoop(cellAction);
 
   return <div class="minesweeper">
-    <GameBoard board={board} dispatch={dispatch} />
-    <Controls startTime={startTime} endTime={endTime} gameStage={gameStage} dispatch={dispatch} />
+    <GameBoard board={gameState.pipe(pluck('board'))} dispatch={dispatch} />
+    <Controls gameState={gameState} dispatch={dispatch} />
   </div>;
 };