+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added JSX/TSX support to RxFM.
 
+### Deprecated
+- Deprecated `DestroySubject`.
+- Deprecated `ConditionalOptions` overload in `conditional` operator.
+
 ## [2.0.0] - 2021-11-26
 
 ### Added